+++ CHANGELOG.md
@@ -1,5 +1,23 @@
+### Release v1.0.6
 ### Bug
+- Prevent searching an empty string & remove onlyLetters [#944](https://github.com/EthVM/EthVM/pull/944)
 - Add Invalid Hash error to exceptions [#933](https://github.com/EthVM/EthVM/pull/933)
+
+### Devop
+
+- Bump ini from 1.3.5 to 1.3.8 [#943](https://github.com/EthVM/EthVM/pull/943)
+- Change Unique chart route names [#942](https://github.com/EthVM/EthVM/pull/942)
+
+===================================================================================
+
+### Release v1.0.5-hotfix.1
+
+### Bug
+
+- Switch tabs for Address Details [#935](https://github.com/EthVM/EthVM/pull/935)
+
+===================================================================================
+
 ### Release v1.0.5
 
 ### Feature
@@ -20,7 +38,6 @@
 - Adjust blocks height to 100% on Address page [#930](https://github.com/EthVM/EthVM/pull/930)
 - Added link to myetherwallet.com to 'powered by mew' image [#928](https://github.com/EthVM/EthVM/pull/928)
 - Fixed incorrect ui layout on fav address page and fav address row mobile [#927](https://github.com/EthVM/EthVM/pull/927)
-
 
 ===================================================================================